pyproject.toml
[tool.poetry]
name = "hybrid-movie-recsys"
version = "0.1.0"
description = "Hybrid Movie Recommendation System with LLM Integration"
authors = ["ThanhNX"]

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.0.0"
transformers = "^4.30.0"
pandas = "^2.0.0"
numpy = "^1.24.0"
scikit-learn = "^1.2.0"
implicit = "^0.7.0"
fastapi = "^0.100.0"
uvicorn = "^0.22.0"
wandb = "^0.15.0"
pyyaml = "^6.0"
python-dotenv = "^1.0.0"
kaggle = "^1.5.13"
pydantic = "^2.0.0"
scipy = "^1.10.0"
tqdm = "^4.65.0"
pyarrow = "^12.0.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
isort = "^5.0.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
